Time: 0 ms (100%), Space: 111.4 MB (11.1%) - LeetHub

+++ 2348-number-of-zero-filled-subarrays/2348-number-of-zero-filled-subarrays.cpp
@@ -3,7 +3,7 @@ public:
     long long zeroFilledSubarray(vector<int>& nums) {
         long long sum = 0;
         long long count = 0;
-        for(long long i = 0; i < nums.size(); i++)
+        for(int i = 0; i < nums.size(); i++)
         {
             if(nums[i] == 0)
             {
